internal cleanup: move the public Rust Window API
Move it from corelib::window::api to corelib::api, where we're going to collect more types in the future
that are meant to be re-exported as public API in the sixtyfps-rs and the interpreter crate.
